requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
smsapi-client
penpyxl
